Migrate crc32 logic to Rust (#55)

+++ ftl/page/page-write.c
@@ -227,5 +227,5 @@ ssize_t page_ftl_write(struct page_ftl *pgftl, struct device_request *request)
 	page_ftl_write_update_metadata(pgftl, paddr, sector);
 	pthread_mutex_unlock(&pgftl->mutex);
 
-	return write_size;
+	return (ssize_t)write_size;
 }
